PelosiBE/requirements.txt
fastapi
uvicorn
pandas
beautifulsoup4
requests
apscheduler
yfinance
python-dotenv
psycopg2-binary
PyJWT
bcrypt